Create findingDescending_order.py
A brute force Algorithm
Input: {11, 13, 18, 15, 9, 4, 23}
output: [five, four, two, three, six, seven, one]in numbers form
dveloping a brute force algorithm to return the descending order of elements in the array
function findingDescending_order(arr):
# variable n to  store the length of the array arr
    n = length of arr
#Create an array order of one element 0
    orders = array of n elements initialized to 0
#Looping variable i
    for i in range(n):
        order = 1
# looping variable j for each element in arr
        for j in range(n):
#If arr[i] is greater arr[i], increment a variable order
            if arr[j] > arr[i]:
                order += 1
#Store value of the order in the orders array
        orders[i] = rank
#Return the orders array
    return orders
#call the function in the output
    findingDescending_order (arr)
----------------------------
Analysing Algorithm:
The worst-case performance for a brute-force algorithm is proportional to the input size. This algorithm works to return the descending order of elements in the array, so, it takes a lot of time because it is very slow even if it is a simple algorithm. it provides an n to store the size of the array, this new array contains one element 0. During the loops over each element, the number of elements in the array will increase under the problem descending order condition. Finally, the algorithm will give or return the order array. The time complexity for this algorithm in the worst case is: O(n^2)
The average case will be as same as the worst case. It gives an estimate of the running time if sorting is random.
The best case: when the array given in the problem is sorted, so, the working will reduce, and the time complexity at the end. O(n)